* Headers/Additions/GNUstepGUI/GSTheme.h, * Source/GSThemeDrawing.m: Correct int/NSInteger mismatch detected by Josh Freeman <pikopixel@twilightedge.com>.

+++ Headers/Additions/GNUstepGUI/GSTheme.h
@@ -1260,7 +1260,7 @@ APPKIT_EXPORT	NSString	*GSThemeWillDeactivateNotification;
 					inView: (NSView *)view
 			      selectingColumns: (BOOL)selectingColumns;
 
-- (void) drawTableViewRow: (int)rowIndex 
+- (void) drawTableViewRow: (NSInteger)rowIndex 
 		 clipRect: (NSRect)clipRect
 		   inView: (NSView *)view;
